scripts/requirements.txt
# Core dependencies
httpx>=0.27.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# PDF extraction (choose one or both)
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Optional: Progress bars
rich>=13.0.0
